// src/features/addresses/address.types.ts
export interface Address {
    id: string;
    no: string;
    addressCode: string;
    officeName: string;
    tel: string;
    fax: string;
    type: string;
    zipCode: string;
    address: string;
    notes: string;
    division: string;
    area: string;
    mainPerson: string;
    branchNumber: string;
    specialNote: string;
    labelName: string;
    labelZip: string;
    labelAddress: string;
    attentionNote: string;
}